% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{checkData}
\usage{
checkData(data = "data.csv", sample.info = "sample.info.csv", path = ".")
}
\arguments{
\item{data}{The name of ms1 data.}

\item{sample.info}{The name of sample information.}

\item{path}{The work directory.}
}
\value{
Check result.
}
\description{
Check data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
