% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compare_rsd}
\alias{compare_rsd}
\title{compare_rsd}
\usage{
compare_rsd(sample.rsd, sample.nor.rsd, QC.rsd, QC.nor.rsd, path = ".")
}
\arguments{
\item{sample.rsd}{sample.rsd}

\item{sample.nor.rsd}{sample.nor.rsd}

\item{QC.rsd}{QC.rsd}

\item{QC.nor.rsd}{QC.nor.rsd}

\item{path}{path}
}
\description{
compare_rsd
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
