% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cvMSE}
\alias{cvMSE}
\title{cvMSE}
\usage{
cvMSE(qc, QC.order, begin1, end1, step1)
}
\arguments{
\item{qc}{qc}

\item{QC.order}{QC.order}

\item{begin1}{begin1}

\item{end1}{end1}

\item{step1}{step1}
}
\description{
cvMSE
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
