% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metNor.R
\name{metNor}
\alias{metNor}
\title{metNor}
\usage{
metNor(
  ms1.data.name = "data.csv",
  sample.info.name = "sample.info.csv",
  minfrac.qc = 0,
  minfrac.sample = 0,
  optimization = TRUE,
  multiple = 5,
  threads = 3,
  path = "."
)
}
\arguments{
\item{ms1.data.name}{MS1 peak table name. Default is "data.csv".}

\item{sample.info.name}{sample.info name. Default is "sample.info.csv"}

\item{minfrac.qc}{Default is 0.}

\item{minfrac.sample}{Default is 0.}

\item{optimization}{Parameter optimization or not?.Default is TRUE.}

\item{multiple}{If multiple = 1, the svr will be built using injection order.
If multiple >= 2, the svr is built using top mutiple peaks correlated peaks.
For tof data, default is 5, for mrm data, default is 1.}

\item{threads}{Number of thread.}

\item{path}{Directory}
}
\description{
Normalize data using different normalization methods.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
