% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metNormalize_logo.R
\name{MetNormalizer_logo}
\alias{MetNormalizer_logo}
\title{MetNormalizer_logo}
\usage{
MetNormalizer_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is `TRUE`
on UTF-8 platforms.}
}
\value{
A ASCII log of MetNormalizer
}
\description{
The MetNormalizer logo, using ASCII or Unicode characters
}
\examples{
MetNormalizer_logo()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
