% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{peak_plot}
\alias{peak_plot}
\title{peak_plot}
\usage{
peak_plot(
  index,
  sample,
  sample.nor,
  QC,
  QC.nor,
  sample.order,
  QC.order,
  tags,
  path = ".",
  sample.rsd = sample.rsd,
  QC.rsd = QC.rsd,
  sample.nor.rsd = sample.nor.rsd,
  QC.nor.rsd = QC.nor.rsd
)
}
\arguments{
\item{index}{index}

\item{sample}{sample.}

\item{sample.nor}{sample.nor.}

\item{QC}{QC}

\item{QC.nor}{QC.nor}

\item{sample.order}{sample.order}

\item{QC.order}{QC.order}

\item{tags}{tags}

\item{path}{path}

\item{sample.rsd}{sample.rsd}

\item{QC.rsd}{QC.rsd}

\item{sample.nor.rsd}{sample.nor.rsd}

\item{QC.nor.rsd}{QC.nor.rsd}
}
\description{
peak_plot
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
