% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTsvrNor.R
\name{svr_function}
\alias{svr_function}
\title{svr_function}
\usage{
svr_function(index, sample, QC, sample.order, QC.order, multiple)
}
\arguments{
\item{index}{index}

\item{sample}{sample}

\item{QC}{QC}

\item{sample.order}{sample.order}

\item{QC.order}{QC.order}

\item{multiple}{multiple}
}
\description{
svr_function
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
