% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTsvrNor.R
\name{SXTsvrNor}
\alias{SXTsvrNor}
\title{SXTsvrNor}
\usage{
SXTsvrNor(
  sample,
  QC,
  tags,
  sample.order,
  QC.order,
  multiple = 5,
  rerun = TRUE,
  peakplot = TRUE,
  path = ".",
  datastyle = "tof",
  dimension1 = TRUE,
  threads = 1
)
}
\arguments{
\item{sample}{sample.}

\item{QC}{QC}

\item{tags}{tags}

\item{sample.order}{sample.order}

\item{QC.order}{QC.order}

\item{multiple}{multiple}

\item{rerun}{rerun}

\item{peakplot}{peakplot}

\item{path}{path}

\item{datastyle}{datastyle}

\item{dimension1}{dimension1}

\item{threads}{threads}
}
\description{
SXTsvrNor.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
